pyproject.toml
[project]
name = "finance-research-agent"
version = "0.1.0"
description = "Google Sheets smoke test utilities for finance research workflows."
readme = "AGENTS.md"
requires-python = ">=3.11"
dependencies = [
  "google-api-python-client==2.168.0",
  "google-auth>=2.47.0,<3.0.0",
  "google-auth-oauthlib==1.2.2",
  "langchain-google-genai>=4.1.3,<5.0.0",
  "langgraph>=1.0.0,<2.0.0",
  "pytest>=8.0.0,<9.0.0",
]

[tool.uv]
package = false
